Tile ID Dictionary
Added tile dictionary to map ID's to tiles

+++ Assets/Scripts/DiscHandler.cs
@@ -21,7 +21,7 @@ public class DiscHandler : MonoBehaviour
     Dictionary<int, Disc> Discs = new Dictionary<int, Disc>();
 
     // Float for making new IDs
-    float ID;
+    int ID;
 
     // ----[Disc Class Goes Here]----
